refactor: rename maintenance folder to plural
Para manter padrão de nomenclatura de páginas, foi mudado nome da página para Maintenances e dentro criado a página Maintenance para representar cada manutenção individual
